feat: add CJS entry point
