+++ index.d.cts
@@ -1,0 +1,7 @@
+/**
+ * Converts [ANSI escape sequences](https://en.wikipedia.org/wiki/ANSI_escape_code)
+ * into human readable text.
+ */
+declare function prettyAnsi(text: string): string;
+
+export = prettyAnsi;
